requirements.txt
aws-cdk.core==1.128.0
aws-cdk.aws-cloud9==1.128.0
aws-cdk.aws-codebuild==1.128.0
aws-cdk.aws-codecommit==1.128.0
aws-cdk.aws-codedeploy==1.128.0
aws-cdk.aws-codepipeline==1.128.0
aws-cdk.aws-codepipeline-actions==1.128.0
aws-cdk.aws-lambda==1.128.0
aws-cdk.aws-lambda-go==1.128.0
aws-cdk.aws-s3==1.128.0
aws-cdk.aws-s3-assets==1.128.0
aws-cdk.aws-s3-deployment==1.128.0
aws-cdk.pipelines==1.128.0
aws-cdk.aws_ssm==1.128.0
black==21.8b0
boto3==1.18.34
botocore==1.21.34
flake8==3.9.2
urllib3==1.26.5
